Refactor: Organize ActivityList unit test files
