+++ rundeckapp/grails-spa/packages/ui-trellis/src/app/components/activity/mocks/setupRundeckContext.ts
